feat: try to run as a cloudflare worker
This is a quick and dirty attempt to run the render function in a
Cloudflare worker. The code is quickly copied from the output of
`npm create cloudflare@latest`. (TS is messy, this would need clean-up.)

After deploying and testing the worker, I ran into Cloudflare's
limitation: `new Function()` is not allowed in workers:
https://developers.cloudflare.com/workers/runtime-apis/web-standards/#javascript-standards

Testing this branch:
$ npx wrangler deploy
$ curl -X POST https://xbrew-render.xbrew.workers.dev \
> -H "Content-Type: application/json" \
> -d '{
>   "code": "import { jsx as _jsx, jsxs as _jsxs } from \"react/jsx-runtime\";\nexport default function Test({ title, content }) {\n    return /*#__PURE__*/ _jsxs(\"div\", {\n        children: [\n            title && /*#__PURE__*/ _jsx(\"h2\", {\n                children: title\n            }),\n            /*#__PURE__*/ _jsx(\"div\", {\n                children: content\n            })\n        ]\n    });\n};",
>   "props": {
>     "title": "Test Title",
>     "content": "Test Content"
>   }
> }'